fix: set desktop mode as the default
- allows double clicking the executable to open the GUI
